Add command to get the version of lib
For Fusion we need to know the exact version of the nativescript-cloud library.
Implement new command that will print the version on the stdout of the process. As we are not able to add additional `options`, use command.
Also improve Gruntfile, so consecutive `grunt test` or `grunt pack` calls will continue to work - ignore the default `.d.ts` file generated by calling `grunt` as it contains references that we do not need.
Improve the speed of the `grunt` by skipping `npm install` and `grunt` in the dependencies in case we've already executed them.