Fix cloud builds by preparing the project before build
Currently we only add the platform before building the application. However, once we try to persist the `.nsbuildinfo` we fail with: `Cannot read property 'changesRequireBuildTime' of null`
The problem is that generating the `.nsbuildinfo` relies on the `.nsprepareinfo`, which is generated during prepare.

So instead of just calling add platform, change the code to call prepare platform. This way the `.nsprepareinfo` will be generated and the project will be prepared.