Create default .gitignore before initial commit
Currently we create the `.gitignore` file right before calling `git push`. However, at this point the initial commit of files is already created and some special files (like `platforms` and `node_modules` directories) are pushed in the remote repository.
Fix this by creating the `.gitignore` at the beginning of the operation, before calling `git add` and `git commit`.
Also rename a method (just to make it clear what's the behavior of the method).