Read cloud env from config files
We can override some of the environment variables on our server by passing special headers. We should provide a way to set the env vars and their values in the headers of the request. The best place is in the server configuration files. The environment variables can be service specific or global.