Fix git builds
When the user X builds project with name test we create local repository with name the sha1 of the project id.
If the user Y creates project with the same name on the same machine we put the new project in the same local repo.
The user Y does not have credentials for the repo of user X and each git operation which requires the credentials fails with 403.
The solution is to include the user email in the repo name.

Also we need to pass the userEmail in our build request properties because on the build machine we need it.