Add cloud build steps
Build steps:
- prepare
- upload
- build
- download

The cloudBuildService will emit buildStepChanged event for each step.
Each object will have build ID, step name and a progress which will be a number between 0 and 100%.

These build steps will be used for profiling the build and some kind of progress indicator in the clients of the extension.