Rename all services names
As each extension may register its own services, it is possible more than one extension to register services with the same name.
This will lead to unexpected behavior at runtime.
Currently it happened with `applicationService` and `gitService` which also exist in `nativescript-starter-kits`.
In order to resolve the issue rename all services by prefixing them with `nsCloud`.

NOTE: This is a breaking change for all publicly exposed services. Here's a map of the breaking changes:

Breaking changes:
* `applicationService` has been renamed to `nsCloudApplicationService`
* `authenticationService` has been renamed to `nsCloudAuthenticationService`
* `cloudBuildService` has been renamed to `nsCloudBuildService`
* `cloudCodesignService` has been renamed to `nsCloudCodesignService`
* `cloudEmulatorLauncher` has been renamed to `nsCloudEmulatorLauncher`
* `cloudPublishService` has been renamed to `nsCloudPublishService`
* `userService` has been renamed to `nsCloudUserService`