Add EULA service and check if EULA is accepted
Add EULA service and check if EULA is accepted before each cloud call. The
EULA must be accepted before executing any command, even before login.
Users should not be able to access cloud infrastructure without accepting
the EULA.

The EULA itself is located on `nativescript.org` - download it from there
once per 24 hours and check if user had accepted it. In case it has not,
do not allow them to access the cloud.

In `user-settings.json` there's a property called `acceptedEulaHash` - it contains the hashsum of the EULA at the moment when the user accepted it.

Rules for checking if user should accept EULA:
1. Check `user-settings.json` - in case there's no property
`acceptedEulaHash` - users must accept EULA
2. In case there's `acceptedEulaHash`, download the EULA from
`nativescript.org`.
2.1 In case the download it successful, compare the shasums of the
downloaded file with the value of the `acceptedEulaHash`. In case they are
different - EULA must be accepted, as it is different from the one the
user had accepted before. In case they are equal, user should not accept
anything and can continue with the work.
2.2 In case the download fail, allow the user to use the tooling - he has
already accepted an EULA, which is most probably the EULA from our site.

There are several optimizations:
- EULA is downloaded only once per process.
- There are two public methods in eula-service - one of them will download
  the EULA only in case it has not been downloaded in the last 24 hours.
  The other one will force the download. It's up to the caller to decide
  which method to use.