Add prompter for accepting EULA on each command
Add `eula-command-helper` that has two methods - one to accept EULA and one to ensure EULA is accepted.
Use the second method in each command that requires cloud operations and in case user had not accepted EULA - prompt him to accept it.
In case the terminal is not interactive or the user does not accept the EULA, command will not be executed.

Add new command - `tns accept eula` - it should be used for automation before executing any other cloud related command. It will not show prompt, it just informs the user that EULA will be accepted and saves the SHA of the current EULA.