Add lockfile for download and for checking current EULA mtime
In case multiple processes try to download the EULA, they should not do it simultaneously as it may cause unexpected failures or incorrect shasum checks.
In order to resolve the issue add locking of the actions by using lockfile on two places - first one is before the check for 24 hours, the other one is on the real download step.