Download EULA to temp file
The EULA should be downloaded to a temp file as the `pipeTo` option of httpClient.httpRequest` immediately creates the file or overwrites it.
In case the download fails, we'll end up with 0KB EULA, which will trigger many incorrect behaviors on our side.
In order to fix this - move the EULA to its location after full download is finished.

Automatically delete tracked files when process ends