Add the application id to the tmp git folder
We need to add the application identifier to the name of the temp git
folder. With the current name we have problem when the user changes the
application identifier of the project. When we create the temp git folder
we set the cloud remote to code commit repo X. After the app identifier is
changed we try to add again cloud remote to code commit repo Y in the same
temp git folder. This operation fails and we default to zip upload.