Check if cloud remote is correct
We need to check if the cloud remote is set correctly before we execute
our git operations. Currently if the remote is set to incorrect url we
will try to add it again and this will fail. We should change the wrong
remote url if the remote is added and if the remote url is incorrect.