Respect proxyAuthenticationRequired property during upload
In case the upload to S3 fails, we dismiss the real error object. However, it may contain additional property: proxyAuthenticationRequired, which shows the caller that the cause of the failure is a proxy, requiring authentication.
So pass the value to the `errors.fail` method, which will construct correct error object.