Add cloud workspace cleanup command
We need to provide our users a way to clean all cloud build data which we
keep for optimizations.

This PR adds command which consumes the new project service which contains
the cleanup endpoint. The endpoint requires the user to be logged in and
to execute the command in the directory of the project which will be
cleaned in the cloud.

The server will clean the code commit data and send messages to all build
machines to clean the incremental build data. In the response we will
receive the Code Commit result and server operation info for each build
machine. We need to wait all server operations to finish and get the
results from all of them.

The `nsCloudProjectService` is added to the public API of the library.