Throw correct error when requests to our server are blocked
In case requests to our server are blocked by proxy, in most of the cases html is returned instead of JSON.
In this case we do not show correct error, but fail with `Unexpected token < in JSON at position 0`
This generates support as users are unable to understand the real cause of the problem.
In order to give them more information, throw error with the returned body.