Fix getting runtime and cli versions for cloud builds
Fix getting runtime and cli versions for cloud builds by reading the
package.json of the project just before sending the request. Currently
we are relying on the nativescriptData passed by the caller, but in case
this is the first build for this platform, the nativeScriptData contains
only the id of the project. However, before sending the build request,
we are adding the platform, so the package.json contains the required
version of the runtime. So instead of using some logic to determine the
runtime version, just read the package.json.
Also remove the default fallbacks for CLI and runtime (3.0.0). From now
on, whenever we are unable to find version that should be used in the
cloud, an error will be raised.
This fixes the case when the first cloud build is executed with version
3.0.x of CLI and runtime, when tns-core-modules version in package.json
is next.
This also fixes the case when first and second cloud build are executed
with different CLI and runtime - when you have CLI 3.4.x, but you are
trying to build a project (for a first time), that has tns-core-modules
3.3.x as dependency, the first cloud build will be executed with CLI
3.3.x and runtime 3.3.x, while the second one will be with 3.4.x

Add tests for versionService.