Add new subscription expiration properties
The current licenseExpiration property represents when the cloud builds
usage will be rest. For the monthly licenses this reset date is the same
as the license expiration. But we don't have only monthly licenses. We
have complimentary licenses with custom expiration date, business and
enterprise licenses which have yearly licenses.
That's why our server now returns 2 new properties in the usage info.
licenseExpirationDate which is the real license expiration date and
resetDate which is the monthly usage reset date. The licenseExpiration
property is not deleted for backwards compatibility.