feat: Improve package.json, so CLI can detect this as extension
In case extension is not installed, but user tries to use any command from it, CLI tries to find out in which extension the command is registered. It works by checking all packages in npm that have the `nativescript:extension` keyword.
After that it reads their package.json and searches for nativescript.commands array.
So prepare the `nativescript-cloud` by adding all commands to the correct place in package.json. Add required keyword and some addtional ones.