fix: Trying to build application with webpack produces incorrect local files
When trying to build the application with webpack, the locally prepared files are incorrect. The problem is that the the bundle value is not passed to the prepare method - we have hardcoded false. This way the local files are prepared without webpack and the result application is not working.
Pass correct value, so the local files will be correct.