Remove validatePlatform for run command
In case when `tns cloud run` command is executed and the local environment is not correctly configured, {N} CLI shows a prompt with 4 options (configure for local builds, configure for cloud builds, configure for both - local and cloud builds, setup manually) to configure automatically the environment. This prompt should not be shown in case when cloud commands are executed.
We discussed the situation and agreed the whole `validatePlatform` method is not needed for cloud run command.