fix: `tns accept eula` command fails
After adding the `If-Modified-Since` flag to http request for downloading EULA, the `tns accept eula` command fails as the server returns empty response and the `$httpClient` fails to `JSON.parse` the result.
So, the command fails as it must download the EULA. In order to fix the issue, force the download in such cases. The case is only one currently - when accepting the EULA. In such case force the download.