Print the publish error
When the cloud publish fails with non ITMS error (e.g. not accepted
license agreement, wrong password), we show only part of the error which
can be `The request could not be completed because:` and the real reason
is hidden. We decided to show the whole stderr to solve this problem.
In case of ITMS error, it will be displayed and the github issues
information displayed by fastlane will be hidden because it is displayed
on the stdout which is filtered.