feat: Remove check for Kinvey EULA when sending requests
In case we send requests to Kinvey, we check if Kinvey EULA has been accepted. This is not required anymore, so remove the check from the `kinvey-request-service`.
As the nsCloudKinveyEulaService has been publicly exposed, we cannot remove it without bumping the major version. So keep it in the code until we prepare for 2.0.0 release