fix: EEXIST error is thrown sometimes when trying to download EULA
In order to ensure EULA is downloaded from one process at a time, we use the `lockfile` package. Currently we use locking on two places:
* Before starting the check if we need to download EULA
* Before executing the actual HTTP call to get the EULA.pdf

The purpose of the first is to ensure we will not execute the check for EULA while other process is currently downloading it. However, the second check also ensures we will not fall in this case. So remove the locking from here.
Also at this point we have a logic to check if the file has been modified in the last 24 hours. In case it has not, we decide that we have to check if there's new EULA. This means that we do not try to download a new EULA only in the first 24 hours after the download.
As this is a complicated logic without real benefit from it, just delete it. We already rely on the `If-Modified-Since` header, so no need for additional safeguards.

Regarding the second locking - move it just before copying the downloaded EULA from its temp location to the real one. This way we'll not create lock in case we do not download a new EULA.
Additionally, ensure the lockfile is unlocked only in case the action for copying the EULA fails. In case the locking operation fails, we must not call unlock as it will delete the lockfile, while other process is using it.
The last change is to conside the lockfile as stale in case it is older than 13 seconds, currently we consider it stale after 10 minutes, but if the EULA cannot be downloaded for 13 seconds, most probably it will not never be downloaded.