fix: hmr is not working when git is used to upload files
Using cloud build with Hot Module Replacement is not working when git is used as it modifies the line endings. Most of the templates are published from Windows machine, so the line endings in them are `\r\n`.
The local prepare of the project produces bundle.js with `hotCurrentHash = <value 1>`. Once we use git to push the files in the cloud, the line endings are `\n`. The `bundle.js` produced in the cloud contains `hotCurrentHash = <value 2>`.
After the built .apk is downloaded from the cloud and installed on device, CLI does not apply full sync as it thinks the local files are the same as the ones in the cloud.
In order to fix this, set `core.autocrlf` to false in the local codecommit repo. This way the line endings will be the same and the `hotCurrentHash` should be the same in the cloud and locally. In order to force creation of new repo, add a versioning to the dir name of the local repo.
This allows us to force creation of new repo in this case.