fix: do not generate nshashes file when --hmr is used
When cloud build is used with hmr, the files generated in the cloud (bundle.js especially) may differ from the local ones.
Currently we build the app in the cloud and generate hashes locally based on the local files, instead of getting them from the cloud. This works in normal case, but not for HMR.
Until we get the hashes from the cloud, skip generation of nshashes in case hmr is passed.