fix: update @types/promise-retry to latest version
Fix the master build by updating `@types/promise-retry` to latest version. Currently the build fails as `@types/promise-retry`'s 1.1.1 version depends on `@types/retry` package with `*`, which means it always installs the latest version.
The `.d.ts` files in the latest `@types/retry` package are for newer version, which has breaking change in the API, so our transpilation fails. The issue is fixed in `@types/promise-retry` by using specific version of the `@types/retry` package.