fix: handle CLI's changes for Ctrl + C
In CLI the `processService` has been removed and the Ctrl + C signal is handled in a different manner. To make the `nativescript-cloud` compatible with both official CLI and the one in developement, introduce process-service in `nativescript-cloud` that wraps the original one, if it exists.
This service should be removed when CLI 5.4.0 is the minimum required one.