fix: publish commands fail to build app
Currently publish commands fail to build the application when `useLegacyWorkflow` is set to false in nsconfig.json. The error happens during project preparation in the cloud with error `The options --release and --hmr cannot be used simultaneously`.
This happens as by default we do not pass `--release` to cloud publish commands, so CLI does not set `options.hmr` to false. So in the cloud we send both `--release` and `--hmr` options and the build fails.
To fix this, set `--no-hmr` when we build in release configuration.