fix: don't call prepare method twice on livesync related commands
CLI 6.0 prepares the project on initial sync for all livesync related commands and the `nativescript-cloud` tries to prepare it again. Hopefully there is a caching based on the platforms for which the webpack compilation is started. This way, a second webpack compilation is not started, but the prepare method is called for the second time. This led to the problem that nativePrepare method is called twice. However, for cloud related operations (when `skipNativePreapare` is true), `prepareNativePlatform` method only checks if there are changes in the project. This way, `checkForChanges` method is called twice. In order to remove the second `checkForChanges` we can remove preparing the project from cloud lib. This will led to the issue that if `tns cloud build ` command is executed and after that `tns cloud run`, the project will be built again. In order to avoid this behavior, we'll force the project preparation on `tns cloud build` command.