fix: tests should be able to run
Currently unit tests are not running as the test-bootstrap requires `errors`, which is using the `deprecated` decorator. The decorator is evaluated when the errors file is required and it expects to have a logger in the globally registered injector.
However, we do not have it in the tests. Remove require of the errors in the test-bootstrap, as it is not needed there.