Snooze log polling until current polling completes
In case the request is a bit slow or the new chunk is huge, getCloudOperationLogs could take a white to complete. In this case, there is no point in polling for logs and it coulkd lead to problems (e.g. if outputCursorPosition is not updated yet, we will print the same logs twice).