Do not generate qrData after cloud build with aab
A user cannot download and install an aab package directly on an Android device as they can with an apk package, so the qrData is irrelevant. Additionally, the originalUrl, returned as part of the qrData is incorrect as it points to an apk file. We rename this apk file as part of the download step of a cloud build operation, so the local file has the correct extension, but the URL is incorrect for the user.