fix: update nativescript-doctor to latest version
The currently used nativescript-doctor version is very old and it has non-direct dependency to the `natives` package, which does not work with latest Node.js versions. It needs to be updated every time when a new Node.js is released.
The latest nativescript-doctor version does not depend on natives package, so the errors with `natives` packages are not propagated anymore.